requirements.txt
Django==2.1.3
django-cors-headers==2.4.0
django-extensions==2.1.3
django-sslserver==0.20
djangorestframework==3.9.0
pytz==2018.7
six==1.11.0
